+++ src/main/java/com/example/ecommerce/modules/order/dto/OrderItemResponseDTO.java
@@ -1,11 +1,13 @@
 package com.example.ecommerce.modules.order.dto;
 
+import java.math.BigDecimal;
+
 public record OrderItemResponseDTO(
         String nomeProduto,
         String skuProduto,
         Integer quantidade,
-        Float precoUnitario,
-        Float subTotal,
+        BigDecimal precoUnitario,
+        BigDecimal subTotal,
         Long productId
 ) {
 }
