+++ src/main/java/com/example/ecommerce/modules/product/dto/ProductResponseResumeDTO.java
@@ -1,12 +1,14 @@
 package com.example.ecommerce.modules.product.dto;
 
+import java.math.BigDecimal;
+
 public record ProductResponseResumeDTO(
         Long id,
         String nome,
         String slug,
         String descricaoCurta,
         String descricao,
-        float preco,
+        BigDecimal preco,
         String sku
 ) {
 }
