+++ src/main/java/com/example/ecommerce/modules/product/dto/ProductUpdateDTO.java
@@ -2,13 +2,15 @@ package com.example.ecommerce.modules.product.dto;
 
 import com.example.ecommerce.modules.product.model.ProductEnum;
 
+import java.math.BigDecimal;
+
 public record ProductUpdateDTO(
         String nome,
         String slug,
         String descricaoCurta,
         String descricao,
-        Float preco,
-        Float precoPromocional,
+        BigDecimal preco,
+        BigDecimal precoPromocional,
         Integer estoqueMinimo,
         Float peso,
         Float altura,
